chore: change folder structure and make icon transparent
