--- src/app/components/CustomIcons/ConnectedSiteIcon.js
@@ -1,46 +1,0 @@
-const ConnectedSiteIcon = (props) => (
-  <svg
-    width={24}
-    height={24}
-    viewBox="0 0 24 24"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-    {...props}
-  >
-    <circle
-      cx={1.57377}
-      cy={1.57377}
-      r={1.57377}
-      transform="matrix(-1 0 0 1 7.14752 5)"
-      fill="currentColor"
-    />
-    <path
-      d="M5.31138 6.33765L8.24908 9.27535"
-      stroke="#374151"
-      strokeWidth={0.9}
-    />
-    <circle cx={18.4262} cy={6.57377} r={1.57377} fill="currentColor" />
-    <path
-      d="M18.7147 6.33765L15.777 9.27535"
-      stroke="#374151"
-      strokeWidth={0.9}
-    />
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M6.21814 16.0572C5.24094 15.6105 4.67273 14.5689 4.88444 13.5155C5.66054 9.65394 8.57524 6.78357 12.0523 6.78357C15.5379 6.78357 18.4583 9.66806 19.2258 13.544C19.4348 14.5994 18.8619 15.6404 17.8815 16.0833C16.1115 16.8828 14.147 17.3279 12.0785 17.3279C9.98771 17.3279 8.00307 16.8731 6.21814 16.0572Z"
-      fill="currentColor"
-    />
-    <path
-      d="M19.2258 13.544L18.7844 13.6314L19.2258 13.544ZM17.8815 16.0833L17.6962 15.6732L17.8815 16.0833ZM6.21814 16.0572L6.03105 16.4665L6.21814 16.0572ZM5.32562 13.6042C6.07142 9.89336 8.84488 7.23357 12.0523 7.23357V6.33357C8.3056 6.33357 5.24966 9.41453 4.44326 13.4269L5.32562 13.6042ZM12.0523 7.23357C15.2675 7.23357 18.0468 9.90654 18.7844 13.6314L19.6672 13.4566C18.8699 9.42959 15.8082 6.33357 12.0523 6.33357V7.23357ZM17.6962 15.6732C15.9833 16.447 14.0818 16.8779 12.0785 16.8779V17.7779C14.2121 17.7779 16.2396 17.3187 18.0667 16.4934L17.6962 15.6732ZM12.0785 16.8779C10.0536 16.8779 8.13262 16.4376 6.40523 15.6479L6.03105 16.4665C7.87352 17.3087 9.92186 17.7779 12.0785 17.7779V16.8779ZM18.7844 13.6314C18.9503 14.4696 18.4974 15.3113 17.6962 15.6732L18.0667 16.4934C19.2265 15.9695 19.9192 14.7292 19.6672 13.4566L18.7844 13.6314ZM4.44326 13.4269C4.18798 14.6971 4.87517 15.9381 6.03105 16.4665L6.40523 15.6479C5.60671 15.2829 5.15748 14.4408 5.32562 13.6042L4.44326 13.4269Z"
-      fill="currentColor"
-    />
-    <path
-      d="M15.186 10.503L11.1653 14.5237L9.1596 12.518"
-      stroke="gray"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    />
-  </svg>
-);
-export default ConnectedSiteIcon;
