+++ src/app/icons/ConnectedSiteIcon.js
@@ -1,0 +1,40 @@
+const ConnectedSiteIcon = (props) => (
+  <svg
+    width={24}
+    height={24}
+    viewBox="0 0 24 24"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+    {...props}
+  >
+    <circle
+      cx="1.57377"
+      cy="1.57377"
+      r="1.57377"
+      transform="matrix(-1 0 0 1 7.14752 5)"
+      fill="currentColor"
+    />
+    <path
+      d="M5.31138 6.33771L8.24908 9.27541"
+      stroke="currentColor"
+      strokeWidth="0.9"
+    />
+    <circle cx="18.4262" cy="6.57377" r="1.57377" fill="currentColor" />
+    <path
+      d="M18.7147 6.33771L15.777 9.27541"
+      stroke="currentColor"
+      strokeWidth="0.9"
+    />
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M4.88444 13.5155C4.67273 14.5689 5.24094 15.6105 6.21814 16.0572C8.00307 16.8731 9.98771 17.3279 12.0785 17.3279C14.147 17.3279 16.1115 16.8828 17.8815 16.0833C18.8619 15.6404 19.4348 14.5994 19.2258 13.544C18.4583 9.66806 15.5379 6.78357 12.0523 6.78357C8.57524 6.78357 5.66054 9.65394 4.88444 13.5155ZM15.8931 11.2101C16.2837 10.8196 16.2837 10.1864 15.8931 9.79588C15.5026 9.40536 14.8694 9.40536 14.4789 9.79588L11.1653 13.1095L9.86671 11.8109C9.47619 11.4204 8.84302 11.4204 8.4525 11.8109C8.06198 12.2014 8.06198 12.8346 8.4525 13.2251L10.4582 15.2308C10.6457 15.4184 10.9001 15.5237 11.1653 15.5237C11.4305 15.5237 11.6849 15.4184 11.8724 15.2308L15.8931 11.2101Z"
+      fill="currentColor"
+    />
+    <path
+      d="M6.21814 16.0572L6.40523 15.6479L6.40523 15.6479L6.21814 16.0572ZM4.88444 13.5155L4.44326 13.4269L4.44326 13.4269L4.88444 13.5155ZM17.8815 16.0833L17.6962 15.6732L17.6962 15.6732L17.8815 16.0833ZM19.2258 13.544L19.6672 13.4566L19.6672 13.4566L19.2258 13.544ZM15.8931 9.79588L15.5749 10.1141L15.5749 10.1141L15.8931 9.79588ZM15.8931 11.2101L15.5749 10.8919L15.5749 10.8919L15.8931 11.2101ZM14.4789 9.79588L14.7971 10.1141L14.7971 10.1141L14.4789 9.79588ZM11.1653 13.1095L10.8471 13.4277L11.1653 13.7459L11.4835 13.4277L11.1653 13.1095ZM9.86671 11.8109L9.54852 12.1291L9.54852 12.1291L9.86671 11.8109ZM8.4525 11.8109L8.7707 12.1291L8.7707 12.1291L8.4525 11.8109ZM8.4525 13.2251L8.7707 12.9069L8.7707 12.9069L8.4525 13.2251ZM10.4582 15.2308L10.7764 14.9126L10.7764 14.9126L10.4582 15.2308ZM11.8724 15.2308L11.5542 14.9126L11.5542 14.9126L11.8724 15.2308ZM6.40523 15.6479C5.60671 15.2829 5.15748 14.4408 5.32562 13.6042L4.44326 13.4269C4.18798 14.6971 4.87517 15.9381 6.03105 16.4665L6.40523 15.6479ZM12.0785 16.8779C10.0536 16.8779 8.13262 16.4376 6.40523 15.6479L6.03105 16.4665C7.87352 17.3087 9.92186 17.7779 12.0785 17.7779V16.8779ZM17.6962 15.6732C15.9833 16.447 14.0818 16.8779 12.0785 16.8779V17.7779C14.2121 17.7779 16.2396 17.3187 18.0667 16.4934L17.6962 15.6732ZM18.7844 13.6314C18.9503 14.4696 18.4974 15.3113 17.6962 15.6732L18.0667 16.4934C19.2265 15.9695 19.9192 14.7292 19.6672 13.4566L18.7844 13.6314ZM12.0523 7.23357C15.2675 7.23357 18.0468 9.90654 18.7844 13.6314L19.6672 13.4566C18.8699 9.42959 15.8082 6.33357 12.0523 6.33357V7.23357ZM5.32562 13.6042C6.07142 9.89336 8.84488 7.23357 12.0523 7.23357V6.33357C8.3056 6.33357 5.24966 9.41453 4.44326 13.4269L5.32562 13.6042ZM15.5749 10.1141C15.7897 10.3289 15.7897 10.6771 15.5749 10.8919L16.2113 11.5283C16.7776 10.962 16.7776 10.0439 16.2113 9.47769L15.5749 10.1141ZM14.7971 10.1141C15.0119 9.89929 15.3601 9.89929 15.5749 10.1141L16.2113 9.47769C15.6451 8.91143 14.727 8.91143 14.1607 9.47769L14.7971 10.1141ZM11.4835 13.4277L14.7971 10.1141L14.1607 9.47769L10.8471 12.7913L11.4835 13.4277ZM9.54852 12.1291L10.8471 13.4277L11.4835 12.7913L10.1849 11.4927L9.54852 12.1291ZM8.7707 12.1291C8.98549 11.9143 9.33373 11.9143 9.54852 12.1291L10.1849 11.4927C9.61865 10.9265 8.70056 10.9265 8.1343 11.4927L8.7707 12.1291ZM8.7707 12.9069C8.55591 12.6921 8.55591 12.3439 8.7707 12.1291L8.1343 11.4927C7.56804 12.059 7.56804 12.9771 8.1343 13.5433L8.7707 12.9069ZM10.7764 14.9126L8.7707 12.9069L8.1343 13.5433L10.14 15.549L10.7764 14.9126ZM11.1653 15.0737C11.0194 15.0737 10.8795 15.0158 10.7764 14.9126L10.14 15.549C10.4119 15.821 10.7807 15.9737 11.1653 15.9737V15.0737ZM11.5542 14.9126C11.4511 15.0158 11.3112 15.0737 11.1653 15.0737V15.9737C11.5499 15.9737 11.9187 15.821 12.1906 15.549L11.5542 14.9126ZM15.5749 10.8919L11.5542 14.9126L12.1906 15.549L16.2113 11.5283L15.5749 10.8919Z"
+      fill="currentColor"
+    />
+  </svg>
+);
+export default ConnectedSiteIcon;
